Update docs to latest meteor/apollo api
I was just setting up a new project and this was a suggestion that came up in the dev tools console. 

I noticed it wasn't up to date here. I have noticed there has been a lot of work in the docs. I am not sure if that update was part of that.